remove duplicate serial number transaction uris - nancy module and client. clarify paged and not paged reducers / selectors.
