--- src/Service.Host/client/src/actions/sernosTransactionsActions.js
@@ -1,11 +1,0 @@
-import { FetchApiActions } from '@linn-it/linn-form-components-library';
-import { sernosTransactionsActionTypes as actionTypes } from './index';
-import * as itemTypes from '../itemTypes';
-import config from '../config';
-
-export default new FetchApiActions(
-    itemTypes.sernosTransactions.actionType,
-    itemTypes.sernosTransactions.uri,
-    actionTypes,
-    config.appRoot
-);
