+++ src/Service.Host/client/src/selectors/serialNumberTransactionsPagedSelectors.js
@@ -1,0 +1,4 @@
+﻿import { PaginationSelectors } from '@linn-it/linn-form-components-library';
+import * as itemTypes from '../itemTypes';
+
+export default new PaginationSelectors(itemTypes.serialNumberTransactionsPaged.item, 'transCode');
