--- src/Service.Host/client/src/selectors/sernosTransactionSelectors.js
@@ -1,4 +1,0 @@
-import { ItemSelectors } from '@linn-it/linn-form-components-library';
-import * as itemTypes from '../itemTypes';
-
-export default new ItemSelectors(itemTypes.sernosTransaction.item);
