--- src/Service.Host/client/src/selectors/sernosTransactionsSelectors.js
@@ -1,4 +1,0 @@
-import { CollectionSelectors } from '@linn-it/linn-form-components-library';
-import * as itemTypes from '../itemTypes';
-
-export default new CollectionSelectors(itemTypes.sernosTransactions.item, 'sernosTransactions');
