--- tests/Integration/Service.Tests/SernosTransactionModuleSpecs/WhenGettingSernosTransaction.cs
@@ -1,49 +1,0 @@
-﻿namespace Linn.Products.Service.Tests.SernosTransactionModuleSpecs
-{
-    using FluentAssertions;
-
-    using Linn.Common.Facade;
-    using Linn.Products.Domain.Linnapps.SernosTransactions;
-    using Linn.Products.Resources;
-
-    using Nancy;
-    using Nancy.Testing;
-
-    using NSubstitute;
-
-    using NUnit.Framework;
-
-    public class WhenGettingSernosTransaction : ContextBase
-    {
-        [SetUp]
-        public void SetUp()
-        {
-            var sernosTrans = new SernosTrans { TransCode = "code", TransDescription = "desc", Comments = "comments" };
-            this.SernosTransactionService.GetById("code")
-                .Returns(new SuccessResult<SernosTrans>(sernosTrans) { Data = sernosTrans });
-
-            this.Response = this.Browser.Get(
-                "/products/maint/sernos-transactions/code",
-                with => { with.Header("Accept", "application/json"); }).Result;
-        }
-
-        [Test]
-        public void ShouldCallService()
-        {
-            this.SernosTransactionService.Received().GetById("code");
-        }
-
-        [Test]
-        public void ShouldReturnOk()
-        {
-            this.Response.StatusCode.Should().Be(HttpStatusCode.OK);
-        }
-
-        [Test]
-        public void ShouldReturnResource()
-        {
-            var resource = this.Response.Body.DeserializeJson<SernosTransactionResource>();
-            resource.TransCode.Should().Be("code");
-        }
-    }
-}
